Csperry

+++ Functions.py
@@ -9,6 +9,7 @@ def helper(filename):
     print(theFile.read())
     #os.startfile('HelpText.txt')
     print("Say exit to return to the main menu")
+
     BacktoMenu = l.listen()
     if BacktoMenu == "exit help" or "exit hell":
         print("closing help manual and returning to main")
@@ -19,8 +20,12 @@ def helper(filename):
 def openFile():
     print("Say File Name to open: <inputFileName>")
     fileString = l.listen() + ".py"
-    fileString.replace(" ", "")
-    #not closing a file here for some reason??
+
+    if os.path.exists(fileString):
+        returnFile = open(fileString, "w")
+        return returnFile.name
+
+
     returnFile = open(fileString, "w+")
     os.chmod(fileString, 0o700)
     returnFile.close()
@@ -35,10 +40,13 @@ def exitMenu(exit):
 def runFile(filename):
     print("Running file....")
 
+
     # runs the file the user has chosen
     if (len(filename) != 0):
         os.system(filename)
 
 
     else:
+
         print("You cannot run a file prior to creating one, call open file and write some code in it first.")
+
